% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{pg_con}
\alias{pg_con}
\title{Create a connection to a database on a PostgreSQL server}
\usage{
pg_con(host_label, dbname, user_label, pw_label, port = "5432")
}
\arguments{
\item{host_label}{a string label to identify the server host in the .Renviron file}

\item{dbname}{a database on a PostgreSQL server}

\item{user_label}{a string label to identify the server username in the .Renviron file}

\item{pw_label}{a string label to identify the server password in the .Renviron file}

\item{port}{a port string used to connect to the database. Defaults to '5432'}
}
\value{
A connection object to a database located on a PostgreSQL server
}
\description{
Create a connection to a database on a PostgreSQL server
}
\examples{
db_con_local = pg_con(host_label = "pg_host_aws",
                      dbname = "sport_sampling",
                      user_label = "pg_me",
                      pw_label = "pg_key",
                      port = "5432")
tbl_count_local = dbGetQuery(db_con_local, paste0("select count(*) from survey_type_lut"))
dbDisconnect(db_con_local)
}
