% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{pg_host}
\alias{pg_host}
\title{Get database server host from .Renviron file}
\usage{
pg_host(host_label)
}
\arguments{
\item{host_label}{the label for the server host in .Renviron}
}
\value{
The host string for the server connection
}
\description{
Get database server host from .Renviron file
}
