% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pg_pw}
\alias{pg_pw}
\title{Get database server password from .Renviron file}
\usage{
pg_pw(pw_label)
}
\arguments{
\item{pw_label}{Alias for password}
}
\description{
Get database server password from .Renviron file
}
