% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{pg_user}
\alias{pg_user}
\title{Get database server username from .Renviron file}
\usage{
pg_user(user_label)
}
\arguments{
\item{user_label}{the label for the server username in .Renviron}
}
\value{
The username string for the server connection
}
\description{
Get database server username from .Renviron file
}
