% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pg_host}
\alias{pg_host}
\title{Get database server host from .Renviron file}
\usage{
pg_host(host_label)
}
\arguments{
\item{Alias}{for host}
}
\description{
Get database server host from .Renviron file
}
