% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{pg_pw}
\alias{pg_pw}
\title{Get database server password from .Renviron file}
\usage{
pg_pw(pw_label)
}
\arguments{
\item{pw_label}{the label for the server password in .Renviron}
}
\value{
The password string for the server connection
}
\description{
Get database server password from .Renviron file
}
