% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pg_user}
\alias{pg_user}
\title{Get database server username from .Renviron file}
\usage{
pg_user(user_label)
}
\arguments{
\item{user_label}{Alias for username}
}
\description{
Get database server username from .Renviron file
}
